refactor: split UI into single-responsibility modules
layout.rs → top_bar.rs, sidebar.rs, fab.rs
notes.rs → note_list.rs, note_card.rs, note_detail.rs
AppState + View → state.rs
Each file = one component, SRP enforced.
