+++ src/ui/note_card.rs
@@ -1,0 +1,56 @@
+use crate::db::Database;
+use crate::models::Note;
+use crate::ui::{AppState, View};
+use dioxus::prelude::*;
+use std::sync::Arc;
+
+#[component]
+pub fn NoteCard(note: Note) -> Element {
+    let mut app: AppState = use_context();
+    let db: Signal<Arc<Database>> = use_context();
+    let note_id = note.id.clone();
+
+    let title = note
+        .title
+        .clone()
+        .unwrap_or_else(|| "Sans titre".to_string());
+
+    let preview = if note.content.len() > 120 {
+        format!("{}...", &note.content[..120])
+    } else {
+        note.content.clone()
+    };
+
+    let date = &note.created_at[..10];
+    let has_audio = note.audio_file_path.is_some();
+
+    let folder_name = db()
+        .folders_for_note(&note.id)
+        .ok()
+        .and_then(|f| f.first().map(|f| f.name.clone()));
+
+    rsx! {
+        div {
+            class: "bg-white p-4 border border-gray-200 rounded-xl mb-2.5",
+            onclick: move |_| {
+                app.view.set(View::NoteDetail { note_id: note_id.clone() });
+            },
+            div { class: "flex justify-between items-center mb-2",
+                h3 { class: "font-semibold text-base text-gray-900", "{title}" }
+                if has_audio {
+                    div { class: "w-2 h-2 rounded-full bg-ios-green" }
+                }
+            }
+            if !preview.is_empty() {
+                p { class: "text-gray-600 text-sm mb-2 line-clamp-2", "{preview}" }
+            }
+            div { class: "flex items-center gap-2",
+                span { class: "text-gray-400 text-xs", "{date}" }
+                if let Some(ref fname) = folder_name {
+                    span { class: "text-xs text-gray-400", "·" }
+                    span { class: "text-xs text-ios-blue", "{fname}" }
+                }
+            }
+        }
+    }
+}
