+++ src/ui/sidebar.rs
@@ -6,52 +6,6 @@ use dioxus::prelude::*;
 use std::sync::Arc;
 
 #[component]
-pub fn TopBar() -> Element {
-    let mut app: AppState = use_context();
-    let db: Signal<Arc<Database>> = use_context();
-    let is_detail = matches!((app.view)(), View::NoteDetail { .. });
-
-    let title = match (app.view)() {
-        View::NotesList => match (app.selected_folder_id)() {
-            Some(ref fid) => db()
-                .get_folder(fid)
-                .ok()
-                .flatten()
-                .map(|f| f.name)
-                .unwrap_or_else(|| "Dossier".to_string()),
-            None => "Toutes mes notes".to_string(),
-        },
-        View::NoteDetail { .. } => "Note".to_string(),
-    };
-
-    rsx! {
-        div { class: "flex items-center px-4 py-3 bg-white border-b border-gray-200 sticky top-0 z-30 gap-3 min-h-[44px]",
-            if is_detail {
-                button {
-                    class: "min-w-[44px] min-h-[44px] flex items-center justify-center text-gray-700",
-                    onclick: move |_| {
-                        app.sliding_out.set(true);
-                        spawn(async move {
-                            futures_timer::Delay::new(std::time::Duration::from_millis(150)).await;
-                            app.sliding_out.set(false);
-                            app.view.set(View::NotesList);
-                        });
-                    },
-                    IconArrowLeft { size: 22 }
-                }
-            } else {
-                button {
-                    class: "min-w-[44px] min-h-[44px] flex items-center justify-center text-gray-700",
-                    onclick: move |_| app.sidebar_open.set(true),
-                    IconList { size: 22 }
-                }
-            }
-            span { class: "text-lg font-semibold text-gray-900 flex-1", "{title}" }
-        }
-    }
-}
-
-#[component]
 pub fn SidebarOverlay() -> Element {
     let mut app: AppState = use_context();
     let is_open = (app.sidebar_open)();
@@ -356,25 +310,6 @@ fn FolderItem(folder: Folder, depth: u32) -> Element {
                     for child in children() {
                         FolderItem { folder: child, depth: depth + 1 }
                     }
-                }
-            }
-        }
-    }
-}
-
-#[component]
-pub fn FloatingActionButton() -> Element {
-    let mut app: AppState = use_context();
-
-    rsx! {
-        div { class: "fixed bottom-6 right-4 z-20",
-            button {
-                class: "w-20 h-20 rounded-full bg-white flex items-center justify-center shadow-xl shadow-gray-400/40",
-                onclick: move |_| {
-                    app.view.set(View::NoteDetail { note_id: String::new() });
-                },
-                div { class: "text-gray-700",
-                    IconNewNote { size: 40 }
                 }
             }
         }
